fix(KEN-816): a restore says what it restored [no-changelog]
`saved` was only ever set on the path where the channel already carried
manifests, so on a fresh channel the recovery test short-circuited and the
run fell through to the message that says what it carried is back in place.
Nothing was put back, because nothing had been saved. `gh` uploads the two
manifests as separate calls, so a run could leave one manifest on the
channel, or a whole release it had just created, and report the prior state
as restored. A half-written channel is bad; one announcing itself as
repaired is what stops anyone looking at it.

The restore now puts the channel back the way the run found it and says
which: the manifests it found go up again, anything it added comes off, and
a channel this run created is deleted. It reads the channel again rather
than assuming, because a failed upload stops partway and only the channel
says how far it got, and a read it cannot make leaves no claim behind. One
`case` on that verdict picks the message, and its default arm names what is
on the channel for whoever has to repair it.

The tag a created channel was made on stays. `gh release create` reuses an
existing tag, so a run cannot tell one it made from one that was already
there carrying no release, and deleting the second would destroy something
it never created. A tag with no release publishes nothing either way.

The guard's own tests move to crates/cli/tests/release_workflow/channel_point.rs.
channel.rs was 25 lines under the test size cap and holds two subjects: which
tags reach the channel, and what the guard does to it. The second is what grew.
Its `gh` stub now keeps the channel as a directory, so a call reads what the
calls before it wrote and the tests can ask what state a run left rather than
which calls it made.
