+++ crates/cli/tests/release_workflow/channel.rs
@@ -173,7 +173,7 @@ fn the_workflow_and_the_build_agree_on_what_a_candidate_is() {
 /// pre-release identifier with a leading zero, an empty identifier, and an
 /// empty build. Published, each would name a manifest version every
 /// candidate then fails to read.
-const REFUSED_VERSIONS: [&str; 6] = [
+pub(crate) const REFUSED_VERSIONS: [&str; 6] = [
     "01.0.0",
     "1.0.0-01",
     "1.0.0-alpha..1",
@@ -202,7 +202,7 @@ fn core_refuses_the_versions_a_regex_would_wave_through() {
 /// What a candidate reading the channel's manifest makes of a version:
 /// `ReleaseFeed::parse` is what runs on it on every machine, and a version
 /// it refuses stops that install from seeing any update at all.
-fn core_can_read(version: &str) -> bool {
+pub(crate) fn core_can_read(version: &str) -> bool {
     let feed = format!(r#"{{"schema":1,"version":"{version}","assets":{{}}}}"#);
     kendex_core::update_feed::ReleaseFeed::parse(feed.as_bytes()).is_ok()
 }
@@ -322,344 +322,13 @@ fn the_channel_is_repointed_for_a_candidate_and_no_other_tag() {
 
 /// One value of the pre-release channel step's `env:` block.
 #[allow(clippy::unwrap_used)]
-fn channel_step_env(name: &str) -> String {
+pub(crate) fn channel_step_env(name: &str) -> String {
     let workflow = workflow();
     step(&workflow, "name: Point the pre-release channel at this tag")
         .iter()
         .find_map(|l| l.trim().strip_prefix(&format!("{name}: ")))
         .unwrap_or_else(|| panic!("the channel step sets no {name}"))
         .to_owned()
-}
-
-/// The guard the channel job runs, as a path this test can execute.
-fn channel_script() -> std::path::PathBuf {
-    std::path::Path::new(env!("CARGO_MANIFEST_DIR")).join(format!("../../{CHANNEL_SCRIPT}"))
-}
-
-/// Named once, and checked below against the line the channel job runs: a
-/// guard renamed on one side and not the other leaves every test here
-/// exercising a file no tag run reaches.
-const CHANNEL_SCRIPT: &str = "tools/release-channel-point";
-
-/// What one run of the channel step did.
-#[cfg(unix)]
-struct Pointed {
-    code: i32,
-    calls: Vec<String>,
-}
-
-#[cfg(unix)]
-impl Pointed {
-    fn ran(&self, verb: &str) -> Option<&String> {
-        self.calls.iter().find(|c| c.starts_with(verb))
-    }
-}
-
-/// The state of the pre-release channel a run finds.
-#[cfg(unix)]
-#[derive(Clone, Copy, Debug)]
-enum Channel<'a> {
-    /// No channel release published yet.
-    Absent,
-    /// Published, with nothing uploaded to it — the one state that leaves
-    /// the guard with no version to compare and lets the write through.
-    Empty,
-    /// Published, its manifest naming this version.
-    Carrying(&'a str),
-    /// Published, with a manifest that names no version at all.
-    Unreadable,
-}
-
-/// Runs the pre-release channel step with `gh` stubbed. `fail` is a
-/// fragment of the one `gh` call that should fail, standing in for the
-/// transient errors every read here has to tell apart from an answer.
-#[cfg(unix)]
-#[allow(clippy::unwrap_used)]
-fn point_channel(channel: Channel, new_version: &str, fail: Option<&str>) -> Pointed {
-    use std::os::unix::fs::PermissionsExt;
-    let dir = tempfile::tempdir().unwrap();
-    let root = rooted(&dir);
-    let dist = root.join("dist");
-    fs::create_dir_all(&dist).unwrap();
-    for name in ["latest.json", "feed.json"] {
-        fs::write(dist.join(name), "{}").unwrap();
-    }
-    let bin = root.join("bin");
-    fs::create_dir_all(&bin).unwrap();
-    let log = root.join("gh.log");
-    // Records every call and answers the three the run reads: whether the
-    // channel is among the repository's releases, whether it carries a
-    // manifest, and what those manifests say. `--dir` and `--pattern` are
-    // honoured rather than assumed, so a step that downloaded somewhere
-    // else, or stopped asking for one of the two manifests, has nothing on
-    // disk to read back or to put back.
-    fs::write(
-        bin.join("gh"),
-        "#!/bin/sh\n\
-         printf '%s\\n' \"$*\" >> \"$GH_LOG\"\n\
-         if [ -n \"$GH_FAIL\" ]; then\n\
-           case \"$*\" in *\"$GH_FAIL\"*) exit 1 ;; esac\n\
-         fi\n\
-         case \"$2\" in\n\
-           */releases)\n\
-             [ -n \"$GH_EXISTS\" ] && printf '%s\\n' \"$CHANNEL\"\n\
-             exit 0\n\
-             ;;\n\
-           */releases/tags/*)\n\
-             [ -n \"$GH_HAS_MANIFEST\" ] && printf 'latest.json\\n'\n\
-             exit 0\n\
-             ;;\n\
-         esac\n\
-         if [ \"$1 $2\" = \"release download\" ]; then\n\
-           while [ $# -gt 0 ]; do\n\
-             [ \"$1\" = \"--dir\" ] && dir=$2\n\
-             [ \"$1\" = \"--pattern\" ] && want=\"$want $2\"\n\
-             shift\n\
-           done\n\
-           [ -n \"$dir\" ] || exit 1\n\
-           mkdir -p \"$dir\"\n\
-           for name in $want; do\n\
-             case \"$name\" in\n\
-               latest.json) printf '%s\\n' \"$GH_MANIFEST\" > \"$dir/latest.json\" ;;\n\
-               feed.json) printf '%s\\n' \"$GH_FEED\" > \"$dir/feed.json\" ;;\n\
-             esac\n\
-           done\n\
-         fi\n\
-         exit 0\n",
-    )
-    .unwrap();
-    fs::set_permissions(bin.join("gh"), fs::Permissions::from_mode(0o755)).unwrap();
-
-    let (exists, has_manifest, manifest) = match channel {
-        Channel::Absent => ("", "", String::new()),
-        Channel::Empty => ("1", "", String::new()),
-        Channel::Carrying(version) => ("1", "1", format!(r#"{{"version":"{version}"}}"#)),
-        Channel::Unreadable => ("1", "1", "{}".to_owned()),
-    };
-
-    // The guard itself, run the way the channel job runs it, rather than a
-    // copy of its text lifted out of the YAML.
-    let run = std::process::Command::new(channel_script())
-        .current_dir(&root)
-        .env_clear()
-        .env(
-            "PATH",
-            format!(
-                "{}:{}",
-                bin.display(),
-                std::env::var("PATH").unwrap_or_default()
-            ),
-        )
-        .env("GH_LOG", &log)
-        .env("GH_EXISTS", exists)
-        .env("GH_HAS_MANIFEST", has_manifest)
-        .env("GH_MANIFEST", &manifest)
-        .env(
-            "GH_FEED",
-            r#"{"schema":1,"version":"published","assets":{}}"#,
-        )
-        .env("GH_FAIL", fail.unwrap_or_default())
-        .env("GITHUB_REPOSITORY", "vanillagreencom/kendex")
-        .env("CHANNEL", channel_step_env("CHANNEL"))
-        .env("NEW_VERSION", new_version)
-        .env("GH_TOKEN", "token")
-        .output()
-        .unwrap();
-    let calls = fs::read_to_string(&log)
-        .unwrap_or_default()
-        .lines()
-        .map(str::to_owned)
-        .collect();
-    Pointed {
-        code: run.status.code().unwrap_or(-1),
-        calls,
-    }
-}
-
-/// A read that failed is not an empty channel. Every one of these leaves
-/// the guard unable to say what the channel carries, and a run that
-/// treated that as "nothing there" would upload over whatever is — which
-/// is the rollback the guard exists to prevent, reached through a lost
-/// connection instead of a stale tag.
-#[cfg(unix)]
-#[test]
-fn a_read_it_could_not_make_stops_the_write() {
-    for (state, failing) in [
-        (Channel::Carrying("2.0.0-rc1"), "/releases --paginate"),
-        (Channel::Carrying("2.0.0-rc1"), "/releases/tags/"),
-        (Channel::Carrying("2.0.0-rc1"), "release download"),
-    ] {
-        let run = point_channel(state, "1.0.0-rc1", Some(failing));
-        assert_ne!(run.code, 0, "{failing} was survivable: {:?}", run.calls);
-        assert!(
-            run.ran("release upload").is_none(),
-            "{failing} still uploaded: {:?}",
-            run.calls
-        );
-    }
-
-    // A manifest that names no version is the same answer: nothing here
-    // can tell whether this tag is ahead of what is published.
-    let run = point_channel(Channel::Unreadable, "1.0.0-rc1", None);
-    assert_ne!(run.code, 0, "an unreadable manifest was survivable");
-    assert!(run.ran("release upload").is_none(), "{:?}", run.calls);
-}
-
-/// A channel version nothing can put in an order is the same answer again.
-/// The guard's whole job is to refuse a write it cannot justify, and it
-/// cannot say `NEW_VERSION` moves the channel forward from a string that is
-/// not a version — so it writes nothing rather than clobbering whatever is
-/// really there. Each of these is refused by core too, asked of core rather
-/// than asserted from the shapes, and the versions the test below drives
-/// through are all accepted, so this cannot pass by refusing everything.
-#[cfg(unix)]
-#[test]
-fn a_channel_version_that_is_not_a_version_stops_the_write() {
-    let malformed: Vec<&str> = REFUSED_VERSIONS
-        .into_iter()
-        .chain(["not a version", "5", "", "v1.0.0", "1.0.0-rc1 "])
-        .collect();
-    for carried in malformed {
-        assert!(
-            !core_can_read(carried),
-            "{carried} reads, so it is not an example of anything"
-        );
-        let run = point_channel(Channel::Carrying(carried), "1.0.0-rc2", None);
-        assert_ne!(run.code, 0, "{carried} was survivable: {:?}", run.calls);
-        assert!(
-            run.ran("release upload").is_none(),
-            "{carried} still uploaded: {:?}",
-            run.calls
-        );
-    }
-}
-
-/// Both manifests reach the channel whether or not the release behind it
-/// exists yet, and the upload replaces what is there. Without `--clobber`
-/// the first candidate's manifests stay and every later one is refused, so
-/// the channel would freeze on rc1 while reporting success.
-#[cfg(unix)]
-#[test]
-fn every_candidate_replaces_both_manifests_on_the_channel() {
-    for state in [
-        Channel::Absent,
-        Channel::Empty,
-        Channel::Carrying("1.0.0-rc1"),
-    ] {
-        let run = point_channel(state, "1.0.0-rc2", None);
-        assert_eq!(run.code, 0, "{state:?}: {:?}", run.calls);
-        assert_eq!(
-            run.ran("release create").is_some(),
-            matches!(state, Channel::Absent),
-            "{state:?}: {:?}",
-            run.calls
-        );
-        let upload = run
-            .ran("release upload")
-            .unwrap_or_else(|| panic!("{state:?} uploaded nothing: {:?}", run.calls));
-        // Named from the URLs rather than by hand: a manifest renamed on
-        // one side and not the other publishes to a name nothing reads.
-        for url in [
-            kendex_core::update_channel::PRERELEASE_FEED_URL,
-            kendex_core::update_channel::PRERELEASE_MANIFEST_URL,
-        ] {
-            let file = url.rsplit('/').next().unwrap_or_default();
-            assert!(
-                upload.contains(&format!("dist/{file}")),
-                "dist/{file} missing from {upload}"
-            );
-        }
-        assert!(upload.contains("--clobber"), "{upload}");
-    }
-}
-
-/// `gh release upload --clobber` deletes an asset before uploading its
-/// replacement, and says in its own help that a failed upload loses the
-/// original. A candidate reading a channel with no `latest.json` sees no
-/// update at all, so an upload that fell over halfway would break every
-/// candidate machine until somebody re-ran a tag. What came down goes back
-/// up. Not an atomic swap — there is no such thing here — but a failure
-/// that leaves the channel as it found it.
-#[cfg(unix)]
-#[test]
-fn a_failed_upload_puts_back_what_the_channel_had() {
-    let run = point_channel(
-        Channel::Carrying("1.0.0-rc1"),
-        "1.0.0-rc2",
-        Some("dist/latest.json"),
-    );
-    assert_ne!(
-        run.code, 0,
-        "a failed upload was survivable: {:?}",
-        run.calls
-    );
-    let restored = run
-        .calls
-        .iter()
-        .filter(|c| c.starts_with("release upload"))
-        .nth(1)
-        .unwrap_or_else(|| panic!("nothing was put back: {:?}", run.calls));
-    // Both names, and from the copies on disk: the shell expanded the glob
-    // against what the download actually wrote, so a manifest that never
-    // came down could not appear here.
-    for name in ["saved/latest.json", "saved/feed.json"] {
-        assert!(restored.contains(name), "{name} not restored: {restored}");
-    }
-    assert!(restored.contains("--clobber"), "{restored}");
-}
-
-/// Whether core reads `latest` as ahead of `running` — the same comparison
-/// the running build makes of the channel it is pointed at, and what the
-/// step in the workflow has to arrive at from bash.
-#[allow(clippy::unwrap_used)]
-fn core_reads_as_newer(latest: &str, running: &str) -> bool {
-    let feed =
-        format!(r#"{{"schema":1,"version":"{latest}","assets":{{"t":"https://example.test/k"}}}}"#);
-    kendex_core::update_feed::ReleaseFeed::parse(feed.as_bytes())
-        .unwrap()
-        .relation_to(running)
-        .unwrap()
-        == kendex_core::update_feed::VersionRelation::Newer
-}
-
-/// The channel only moves forward. A tag re-run after a later one would
-/// otherwise put its own older manifests back, and every candidate machine
-/// would quietly drop to that release and stop being offered anything —
-/// which on this channel is indistinguishable from the update path being
-/// broken, the very thing the channel exists to test.
-///
-/// The step's bash and core have to reach one answer, so the expectation
-/// is asked of core rather than written out: whatever a build would call
-/// newer is what the channel is allowed to move to.
-#[cfg(unix)]
-#[test]
-fn a_tag_behind_the_channel_leaves_it_alone() {
-    let versions = [
-        "1.0.0-rc1",
-        "1.0.0-rc2",
-        "1.0.0-rc9",
-        "1.0.0-rc10",
-        "1.0.0-beta.1",
-        "1.1.0-rc1",
-        "2.0.0-rc1",
-        "1.0.0-rc2+build",
-    ];
-    for carried in versions {
-        for tagged in versions {
-            let run = point_channel(Channel::Carrying(carried), tagged, None);
-            assert_eq!(run.code, 0, "{carried} -> {tagged}: {:?}", run.calls);
-            // Equal versions re-run the same tag, which refreshes rather
-            // than rolls back, so only a strictly newer carried version
-            // holds the channel.
-            let refused = core_reads_as_newer(carried, tagged);
-            assert_eq!(
-                run.ran("release upload").is_none(),
-                refused,
-                "channel carries {carried}, tag is {tagged}: {:?}",
-                run.calls
-            );
-        }
-    }
 }
 
 /// What a burst of tags does to one job, under GitHub's concurrency rules:
@@ -732,27 +401,6 @@ fn the_channel_write_is_the_only_thing_a_group_holds() {
             .iter()
             .any(|l| l.trim() == "cancel-in-progress: false"),
         "the {writing} job cancels a channel write in progress"
-    );
-}
-
-/// The guard is a file the workflow names, so the two have to agree: renamed
-/// on one side, every test here exercises something no tag run reaches.
-#[test]
-fn the_channel_job_runs_the_guard_these_tests_run() {
-    let workflow = workflow();
-    assert!(
-        step(&workflow, "name: Point the pre-release channel at this tag")
-            .iter()
-            .any(|l| l.trim() == format!("run: {CHANNEL_SCRIPT}")),
-        "the channel step does not run {CHANNEL_SCRIPT}"
-    );
-    assert!(channel_script().is_file(), "{CHANNEL_SCRIPT} is not a file");
-    // The job has to check the repository out, or that line runs nothing.
-    assert!(
-        job(&workflow, job_declaring(&workflow, CHANNEL_SCRIPT))
-            .iter()
-            .any(|l| l.contains("actions/checkout")),
-        "the channel job never checks out {CHANNEL_SCRIPT}"
     );
 }
