+++ crates/cli/tests/release_workflow/main.rs
@@ -664,4 +664,5 @@ fn no_lane_triple_is_hardcoded_into_build_or_staging() {
 }
 
 mod channel;
+mod channel_point;
 mod signing;
